Create an AWS Batch monitor object.

Create an R6 object to manage AWS Batch jobs and job definitions.

crew_aws_batch_monitor(
  job_queue,
  job_definition = paste0("crew-aws-batch-job-definition-", crew::crew_random_name()),
  log_group = "/aws/batch/job",
  config = NULL,
  credentials = NULL,
  endpoint = NULL,
  region = NULL
)

Arguments

job_queue

Character of length 1, name of the AWS Batch job queue.

job_definition

Character of length 1, name of the AWS Batch job definition. The job definition might or might not exist at the time crew_aws_batch_monitor() is called. Either way is fine.

log_group

Character of length 1, AWS Batch CloudWatch log group to get job logs. The default log group is often "/aws/batch/job", but not always. It is not easy to get the log group of an active job or job definition, so if you have a non-default log group and you do not know its name, please consult your system administrator.

config

Optional named list, config argument of paws.compute::batch() with optional configuration details.

credentials

Optional named list. credentials argument of paws.compute::batch() with optional credentials (if not already provided through environment variables such as AWS_ACCESS_KEY_ID).

endpoint

Optional character of length 1. endpoint argument of paws.compute::batch() with the endpoint to send HTTP requests.

region

Character of length 1. region argument of paws.compute::batch() with an AWS region string such as "us-east-2". Serves as the region for both AWS Batch and CloudWatch. Tries to default to paws.common::get_config()$region, then to Sys.getenv("AWS_REGION") if unsuccessful, then Sys.getenv("AWS_REGION"), then Sys.getenv("AWS_DEFAULT_REGION").

See also