AWS Batch launcher class

AWS Batch launcher R6 class

Details

See crew_launcher_aws_batch().

See also

Super class

crew::crew_class_launcher -> crew_class_launcher_aws_batch

Active bindings

aws_batch_config

See crew_launcher_aws_batch().

aws_batch_credentials

See crew_launcher_aws_batch().

aws_batch_endpoint

See crew_launcher_aws_batch().

aws_batch_region

See crew_launcher_aws_batch().

aws_batch_job_definition

See crew_launcher_aws_batch().

aws_batch_job_queue

See crew_launcher_aws_batch().

aws_batch_share_identifier

See crew_launcher_aws_batch().

aws_batch_scheduling_priority_override

See crew_launcher_aws_batch().

aws_batch_parameters

See crew_launcher_aws_batch().

aws_batch_container_overrides

See crew_launcher_aws_batch().

aws_batch_node_overrides

See crew_launcher_aws_batch().

aws_batch_retry_strategy

See crew_launcher_aws_batch().

aws_batch_propagate_tags

See crew_launcher_aws_batch().

aws_batch_timeout

See crew_launcher_aws_batch().

aws_batch_tags

See crew_launcher_aws_batch().

aws_batch_eks_properties_override

See crew_launcher_aws_batch().

Methods

Public methods

crew_class_launcher_aws_batch$new()

crew_class_launcher_aws_batch$validate()

crew_class_launcher_aws_batch$args_client()

crew_class_launcher_aws_batch$args_submit()

crew_class_launcher_aws_batch$launch_worker()

crew_class_launcher_aws_batch$terminate_worker()

Inherited methods

crew::crew_class_launcher$active()

crew::crew_class_launcher$booting()

crew::crew_class_launcher$call()

crew::crew_class_launcher$done()

crew::crew_class_launcher$errors()

crew::crew_class_launcher$forward()

crew::crew_class_launcher$launch()

crew::crew_class_launcher$rotate()

crew::crew_class_launcher$scale()

crew::crew_class_launcher$set_name()

crew::crew_class_launcher$settings()

crew::crew_class_launcher$start()

crew::crew_class_launcher$summary()

crew::crew_class_launcher$tally()

crew::crew_class_launcher$terminate()

crew::crew_class_launcher$terminate_workers()

crew::crew_class_launcher$unlaunched()

crew::crew_class_launcher$wait()

Method new()

Abstract launcher constructor.

Usage

crew_class_launcher_aws_batch$new(
  name = NULL,
  seconds_interval = NULL,
  seconds_timeout = NULL,
  seconds_launch = NULL,
  seconds_idle = NULL,
  seconds_wall = NULL,
  tasks_max = NULL,
  tasks_timers = NULL,
  reset_globals = NULL,
  reset_packages = NULL,
  reset_options = NULL,
  garbage_collection = NULL,
  launch_max = NULL,
  tls = NULL,
  processes = NULL,
  aws_batch_config = NULL,
  aws_batch_credentials = NULL,
  aws_batch_endpoint = NULL,
  aws_batch_region = NULL,
  aws_batch_job_definition = NULL,
  aws_batch_job_queue = NULL,
  aws_batch_share_identifier = NULL,
  aws_batch_scheduling_priority_override = NULL,
  aws_batch_parameters = NULL,
  aws_batch_container_overrides = NULL,
  aws_batch_node_overrides = NULL,
  aws_batch_retry_strategy = NULL,
  aws_batch_propagate_tags = NULL,
  aws_batch_timeout = NULL,
  aws_batch_tags = NULL,
  aws_batch_eks_properties_override = NULL
)

Arguments

name

See crew_launcher_aws_batch().

seconds_interval

See crew_launcher_aws_batch().

seconds_timeout

See crew_launcher_aws_batch().

seconds_launch

See crew_launcher_aws_batch().

seconds_idle

See crew_launcher_aws_batch().

seconds_wall

See crew_launcher_aws_batch().

tasks_max

See crew_launcher_aws_batch().

tasks_timers

See crew_launcher_aws_batch().

reset_globals

See crew_launcher_aws_batch().

reset_packages

See crew_launcher_aws_batch().

reset_options

See crew_launcher_aws_batch().

garbage_collection

See crew_launcher_aws_batch().

launch_max

See crew_launcher_aws_batch().

tls

See crew_launcher_aws_batch().

processes

See crew_launcher_aws_batch().

aws_batch_config

See crew_launcher_aws_batch().

aws_batch_credentials

See crew_launcher_aws_batch().

aws_batch_endpoint

See crew_launcher_aws_batch().

aws_batch_region

See crew_launcher_aws_batch().

aws_batch_job_definition

See crew_launcher_aws_batch().

aws_batch_job_queue

See crew_launcher_aws_batch().

aws_batch_share_identifier

See crew_launcher_aws_batch().

aws_batch_scheduling_priority_override

See crew_launcher_aws_batch().

aws_batch_parameters

See crew_launcher_aws_batch().

aws_batch_container_overrides

See crew_launcher_aws_batch().

aws_batch_node_overrides

See crew_launcher_aws_batch().

aws_batch_retry_strategy

See crew_launcher_aws_batch().

aws_batch_propagate_tags

See crew_launcher_aws_batch().

aws_batch_timeout

See crew_launcher_aws_batch().

aws_batch_tags

See crew_launcher_aws_batch().

aws_batch_eks_properties_override

See crew_launcher_aws_batch().

Returns

An abstract launcher object.

Method validate()

Validate the launcher.

Usage

crew_class_launcher_aws_batch$validate()

Returns

NULL (invisibly). Throws an error if a field is invalid.

Method args_client()

Argument list for paws.compute::batch().

Usage

crew_class_launcher_aws_batch$args_client()

Returns

Argument list for paws.compute::batch().

Method args_submit()

Argument list for paws.compute::batch()$submit_job().

Usage

crew_class_launcher_aws_batch$args_submit(call, name)

Arguments

call

Character of length 1, a namespaced call to crew_worker() which will run in the worker and accept tasks.

name

Character of length 1, an informative worker name.

Returns

Argument list for paws.compute::batch()$submit_job().

Method launch_worker()

Launch a local process worker which will dial into a socket.

Usage

crew_class_launcher_aws_batch$launch_worker(
  call,
  name,
  launcher,
  worker,
  instance
)

Arguments

call

Character of length 1, a namespaced call to crew_worker() which will run in the worker and accept tasks.

name

Character of length 1, an informative worker name.

launcher

Character of length 1, name of the launcher.

worker

Positive integer of length 1, index of the worker. This worker index remains the same even when the current instance of the worker exits and a new instance launches. It is always between 1 and the maximum number of concurrent workers.

instance

Character of length 1 to uniquely identify the current instance of the worker.

Details

The call argument is R code that will run to initiate the worker.

Returns

A handle object to allow the termination of the worker later on.

Method terminate_worker()

Terminate a local process worker.

Usage

crew_class_launcher_aws_batch$terminate_worker(handle)

Arguments

handle

A process handle object previously returned by launch_worker().

Returns

NULL (invisibly).