AWS Batch monitor class

AWS Batch monitor R6 class

Details

See crew_monitor_aws_batch().

See also

Active bindings

job_queue

See crew_monitor_aws_batch().

job_definition

See crew_monitor_aws_batch().

log_group

See crew_monitor_aws_batch().

config

See crew_monitor_aws_batch().

credentials

See crew_monitor_aws_batch().

endpoint

See crew_monitor_aws_batch().

region

See crew_monitor_aws_batch().

Methods

Public methods

crew_class_monitor_aws_batch$new()

crew_class_monitor_aws_batch$validate()

crew_class_monitor_aws_batch$register()

crew_class_monitor_aws_batch$deregister()

crew_class_monitor_aws_batch$describe()

crew_class_monitor_aws_batch$submit()

crew_class_monitor_aws_batch$terminate()

crew_class_monitor_aws_batch$status()

crew_class_monitor_aws_batch$log()

crew_class_monitor_aws_batch$jobs()

crew_class_monitor_aws_batch$active()

crew_class_monitor_aws_batch$inactive()

crew_class_monitor_aws_batch$submitted()

crew_class_monitor_aws_batch$pending()

crew_class_monitor_aws_batch$runnable()

crew_class_monitor_aws_batch$starting()

crew_class_monitor_aws_batch$running()

crew_class_monitor_aws_batch$succeeded()

crew_class_monitor_aws_batch$failed()

Method new()

AWS Batch job definition constructor.

Usage

crew_class_monitor_aws_batch$new(
  job_queue = NULL,
  job_definition = NULL,
  log_group = NULL,
  config = NULL,
  credentials = NULL,
  endpoint = NULL,
  region = NULL
)

Arguments

job_queue

See crew_monitor_aws_batch().

job_definition

See crew_monitor_aws_batch().

log_group

See crew_monitor_aws_batch().

config

See crew_monitor_aws_batch().

credentials

See crew_monitor_aws_batch().

endpoint

See crew_monitor_aws_batch().

region

See crew_monitor_aws_batch().

Returns

AWS Batch job definition object.

Method validate()

Validate the object.

Usage

crew_class_monitor_aws_batch$validate()

Returns

NULL (invisibly). Throws an error if a field is invalid.

Method register()

Register a job definition.

Usage

crew_class_monitor_aws_batch$register(
  image,
  platform_capabilities = "EC2",
  memory_units = "gigabytes",
  memory = NULL,
  cpus = NULL,
  gpus = NULL,
  seconds_timeout = NULL,
  scheduling_priority = NULL,
  tags = NULL,
  propagate_tags = NULL,
  parameters = NULL,
  job_role_arn = NULL,
  execution_role_arn = NULL
)

Arguments

image

Character of length 1, Docker image used for each job. You can supply a path to an image in Docker Hub or the full URI of an image in an Amazon ECR repository.

platform_capabilities

Optional character of length 1, either "EC2" to run on EC2 or "FARGATE" to run on Fargate.

memory_units

Character of length 1, either "gigabytes" or "mebibytes" to set the units of the memory argument. "gigabytes" is simpler for EC2 jobs, but Fargate has strict requirements about specifying exact amounts of mebibytes (MiB). for details, read https://docs.aws.amazon.com/cli/latest/reference/batch/register-job-definition.html # nolint

memory

Positive numeric of length 1, amount of memory to request for each job.

cpus

Positive numeric of length 1, number of virtual CPUs to request for each job.

gpus

Positive numeric of length 1, number of GPUs to request for each job.

seconds_timeout

Optional positive numeric of length 1, number of seconds until a job times out.

scheduling_priority

Optional nonnegative integer of length 1 between 0 and 9999, priority of jobs. Jobs with higher-valued priorities are scheduled first. The priority only applies if the job queue has a fair share policy. Set to NULL to omit.

tags

Optional character vector of tags.

propagate_tags

Optional logical of length 1, whether to propagate tags from the job or definition to the ECS task.

parameters

Optional character vector of key-value pairs designating parameters for job submission.

job_role_arn

Character of length 1, Amazon resource name (ARN) of the job role.

execution_role_arn

Character of length 1, Amazon resource name (ARN) of the execution role.

Details

The register() method registers a simple job definition using the job definition name and log group originally supplied to crew_monitor_aws_batch(). Job definitions created with $register() are container-based and use the AWS log driver. For more complicated kinds of jobs, we recommend skipping register(): first call https://www.paws-r-sdk.com/docs/batch_register_job_definition/ to register the job definition, then supply the job definition name to the job_definition argument of crew_monitor_aws_batch().

Returns

A one-row tibble with the job definition name, ARN, and revision number of the registered job definition.

Method deregister()

Attempt to deregister the job definition.

Usage

crew_class_monitor_aws_batch$deregister()

Details

Attempt to deregister the job definition whose name was originally supplied to the job_definition argument of crew_monitor_aws_batch().

Returns

NULL (invisibly).

Method describe()

Describe the current active revision of the job definition.

Usage

crew_class_monitor_aws_batch$describe()

Returns

If the job definition is not active or does not exist, describe() returns NULL. Otherwise, it returns a tibble with job definition information. Some fields may be nested lists.

Method submit()

Submit a single AWS Batch job to the given job queue under the given job definition.

Usage

crew_class_monitor_aws_batch$submit(
  command = c("sleep", "300"),
  name = paste0("crew-aws-batch-job-", crew::crew_random_name()),
  memory_units = "gigabytes",
  memory = NULL,
  cpus = NULL,
  gpus = NULL,
  seconds_timeout = NULL,
  share_identifier = NULL,
  scheduling_priority_override = NULL,
  tags = NULL,
  propagate_tags = NULL,
  parameters = NULL
)

Arguments

command

Character vector with the command to submit for the job. Usually a Linux shell command with each term in its own character string.

name

Character of length 1 with the job name.

memory_units

Character of length 1, either "gigabytes" or "mebibytes" to set the units of the memory argument. "gigabytes" is simpler for EC2 jobs, but Fargate has strict requirements about specifying exact amounts of mebibytes (MiB). for details, read https://docs.aws.amazon.com/cli/latest/reference/batch/register-job-definition.html # nolint

memory

Positive numeric of length 1, amount of memory to request for each job.

cpus

Positive numeric of length 1, number of virtual CPUs to request for each job.

gpus

Positive numeric of length 1, number of GPUs to request for each job.

seconds_timeout

Optional positive numeric of length 1, number of seconds until a job times out.

share_identifier

Character of length 1 with the share identifier of the job. Only applies if the job queue has a scheduling policy. Read the official AWS Batch documentation for details.

scheduling_priority_override

Optional nonnegative integer of length between 0 and 9999, priority of the job. This value overrides the priority in the job definition. Jobs with higher-valued priorities are scheduled first. The priority applies if the job queue has a fair share policy. Set to NULL to omit.

tags

Optional character vector of tags.

propagate_tags

Optional logical of length 1, whether to propagate tags from the job or definition to the ECS task.

parameters

Optional character vector of key-value pairs designating parameters for job submission.

Details

This method uses the job queue and job definition that were supplied through crew_monitor_aws_batch(). Any jobs submitted this way are different from the crew workers that the crew controller starts automatically using the AWS Batch launcher plugin. You may use the submit() method in the monitor for different purposes such as testing.

Returns

A one-row tibble with the name, ID, and Amazon resource name (ARN) of the job.

Method terminate()

Terminate one or more AWS Batch jobs.

Usage

crew_class_monitor_aws_batch$terminate(
  ids,
  reason = "terminated by crew.aws.batch monitor",
  verbose = TRUE
)

Arguments

ids

Character vector with the IDs of the AWS Batch jobs to terminate.

reason

Character of length 1, natural language explaining the reason the job was terminated.

verbose

Logical of length 1, whether to show a progress bar if the R process is interactive and length(ids) is greater than 1.

Returns

NULL (invisibly).

Method status()

Get the status of a single job

Usage

crew_class_monitor_aws_batch$status(id)

Arguments

id

Character of length 1, job ID. This is different from the user-supplied job name.

Returns

A one-row tibble with information about the job.

Method log()

Get the CloudWatch log of a job.

Usage

crew_class_monitor_aws_batch$log(id, start_from_head = FALSE)

Arguments

id

Character of length 1, job ID. This is different from the user-supplied job name.

start_from_head

Logical of length 1, whether to print earlier log events before later ones.

Details

This method assumes the job has log driver "awslogs" (specifying AWS CloudWatch) and that the log group is the one prespecified in the log_group argument of crew_monitor_aws_batch(). This method cannot use other log drivers such as Splunk, and it will fail if the log group is wrong or missing.

Returns

A tibble with log information.

Method jobs()

List all the jobs in the given job queue with the given job definition.

Usage

crew_class_monitor_aws_batch$jobs(
  status = c("submitted", "pending", "runnable", "starting", "running", "succeeded",
    "failed")
)

Arguments

status

Character vector of job states. Results are limited to these job states.

Details

The output only includes jobs under the job queue and job definition that were supplied through crew_monitor_aws_batch().

Returns

A tibble with one row per job and columns with job information.

Method active()

List active jobs: submitted, pending, runnable, starting, or running (not succeeded or failed).

Usage

crew_class_monitor_aws_batch$active()

Details

The output only includes jobs under the job queue and job definition that were supplied through crew_monitor_aws_batch().

Returns

A tibble with one row per job and columns with job information.

Method inactive()

List inactive jobs: ones whose status is succeeded or failed (not submitted, pending, runnable, starting, or running).

Usage

crew_class_monitor_aws_batch$inactive()

Details

The output only includes jobs under the job queue and job definition that were supplied through crew_monitor_aws_batch().

Returns

A tibble with one row per job and columns with job information.

Method submitted()

List jobs whose status is "submitted".

Usage

crew_class_monitor_aws_batch$submitted()

Details

The output only includes jobs under the job queue and job definition that were supplied through crew_monitor_aws_batch().

Returns

A tibble with one row per job and columns with job information.

Method pending()

List jobs whose status is "pending".

Usage

crew_class_monitor_aws_batch$pending()

Details

The output only includes jobs under the job queue and job definition that were supplied through crew_monitor_aws_batch().

Returns

A tibble with one row per job and columns with job information.

Method runnable()

List jobs whose status is "runnable".

Usage

crew_class_monitor_aws_batch$runnable()

Details

The output only includes jobs under the job queue and job definition that were supplied through crew_monitor_aws_batch().

Returns

A tibble with one row per job and columns with job information.

Method starting()

List jobs whose status is "starting".

Usage

crew_class_monitor_aws_batch$starting()

Details

The output only includes jobs under the job queue and job definition that were supplied through crew_monitor_aws_batch().

Returns

A tibble with one row per job and columns with job information.

Method running()

List jobs whose status is "running".

Usage

crew_class_monitor_aws_batch$running()

Details

The output only includes jobs under the job queue and job definition that were supplied through crew_monitor_aws_batch().

Returns

A tibble with one row per job and columns with job information.

Method succeeded()

List jobs whose status is "succeeded".

Usage

crew_class_monitor_aws_batch$succeeded()

Details

The output only includes jobs under the job queue and job definition that were supplied through crew_monitor_aws_batch().

Returns

A tibble with one row per job and columns with job information.

Method failed()

List jobs whose status is "failed".

Usage

crew_class_monitor_aws_batch$failed()

Details

The output only includes jobs under the job queue and job definition that were supplied through crew_monitor_aws_batch().

Returns

A tibble with one row per job and columns with job information.